+++ onenote_export/page_run_pandoc_conversion_to_markdown.py
@@ -1,0 +1,5 @@
+from onenote_export.OneNotePageExportTaskContext import OneNotePageExportTaskContext
+
+
+def page_run_pandoc_conversion_to_markdown(context: OneNotePageExportTaskContext):
+    context.run_pandoc_conversion_to_markdown()
