+++ test_onenote_export/TestOneNotePageExporter.py
@@ -5,6 +5,7 @@ from unittest.mock import MagicMock, patch, mock_open
 
 from onenote import OneNotePage, OneNoteSection, OneNoteSectionGroup, OneNoteNotebook, OneNoteApplication
 from onenote_export.OneNoteExportTaskContext import OneNoteExportTaskContext
+from onenote_export.OneNoteExportTaskFactory import OneNoteExportTaskFactory
 from onenote_export.OneNotePageExportTaskContext import OneNotePageExportTaskContext
 from onenote_export.OneNotePageExporter import OneNotePageExporter
 from onenote_export.TemporaryOneNotePageDocxExport import TemporaryOneNotePageDocxExport
@@ -15,7 +16,7 @@ from path_scrubbing import PathComponentScrubber
 class TestOneNotePageExporter(unittest.TestCase):
     def test_can_instantiate(self):
         # Arrange
-        subject_ctor_args = (self._mock_context(), ())
+        subject_ctor_args = (self._mock_context(), (), MagicMock(spec=OneNoteExportTaskFactory))
         subject_ctor_kwargs = {}
 
         # Act
@@ -84,8 +85,13 @@ class TestOneNotePageExporter(unittest.TestCase):
         return page_node
 
     def _create_subject_with_mocked_dependencies(self) -> OneNotePageExporter:
-        subject_ctor_args = ()
-        subject_ctor_kwargs = {'context': self._mock_context(), 'prerequisites': ()}
+        subject_ctor_args = (
+        )
+        subject_ctor_kwargs = {
+            'context': self._mock_context(),
+            'prerequisites': (),
+            'subtask_factory': MagicMock(spec=OneNoteExportTaskFactory)
+        }
 
         return OneNotePageExporter(*subject_ctor_args, **subject_ctor_kwargs)
